+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 3.10.0
+* BREAKING: Removed the context parameter for `MenuButton` `onPressed`. If you need the menu button context specifically, use `onContextPressed` in the `MenuButton` instead. Both callbacks will fire if both are defined.
 * BREAKING: Removed `Adaptive`
 * BREAKING: Removed `AbstractArcaneTheme`
 * BREAKING: Removed `FillScreen` & `SliverScreen` `minContentFraction` & `minContentWidth`